[add] isoinstant logical type.

+++ lang/java/avro-joda/avro-joda-ext/src/main/java/org/apache/avro/joda/IsoInstant.java
@@ -56,7 +56,7 @@ public class IsoInstant extends AbstractLogicalType {
 
   @Override
   public Object deserialize(Object object) {
-    return new Instant(object);
+    return new Instant((Long) object);
   }
 
   @Override
